pyproject.toml
[tool.poetry]
name = "electronics-network"
version = "0.1.0"
description = ""
authors = ["BroneKot"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
python-dotenv = "^1.0.1"
coverage = "^7.6.1"
isort = "^5.13.2"
gunicorn = "^23.0.0"
flake8 = "^7.1.1"
whitenoise = "^6.7.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
